Changes in Homework_8 acc. to feedback

+++ src/Homework_8/Square.java
@@ -1,7 +1,7 @@
 package Homework_8;
 
 public class Square extends Quadrangle {
-    public Square(int side) {
+    public Square(double side) {
         length = side;
         width = side;
     }
